+++ extension/src/nmost/max_int.rs
@@ -3,6 +3,7 @@ use pgx::{iter::SetOfIterator, *};
 use crate::nmost::*;
 
 use crate::{
+    accessors::{AccessorIntoArray, AccessorIntoValues},
     flatten,
     palloc::{Inner, Internal, InternalAsValue, ToInternal},
     pg_type,
@@ -14,40 +15,35 @@ use std::cmp::Reverse;
 
 type MaxIntTransType = NMostTransState<Reverse<i64>>;
 
-#[pg_schema]
-pub mod toolkit_experimental {
-    use super::*;
-
-    pg_type! {
-        #[derive(Debug)]
-        struct MaxInts <'input> {
-            capacity : u32,
-            elements : u32,
-            values : [i64; self.elements],
-        }
+pg_type! {
+    #[derive(Debug)]
+    struct MaxInts <'input> {
+        capacity : u32,
+        elements : u32,
+        values : [i64; self.elements],
     }
-    ron_inout_funcs!(MaxInts);
+}
+ron_inout_funcs!(MaxInts);
 
-    impl<'input> From<&mut MaxIntTransType> for MaxInts<'input> {
-        fn from(item: &mut MaxIntTransType) -> Self {
-            let heap = std::mem::take(&mut item.heap);
-            unsafe {
-                flatten!(MaxInts {
-                    capacity: item.capacity as u32,
-                    elements: heap.len() as u32,
-                    values: heap
-                        .into_sorted_vec()
-                        .into_iter()
-                        .map(|x| x.0)
-                        .collect::<Vec<i64>>()
-                        .into()
-                })
-            }
+impl<'input> From<&mut MaxIntTransType> for MaxInts<'input> {
+    fn from(item: &mut MaxIntTransType) -> Self {
+        let heap = std::mem::take(&mut item.heap);
+        unsafe {
+            flatten!(MaxInts {
+                capacity: item.capacity as u32,
+                elements: heap.len() as u32,
+                values: heap
+                    .into_sorted_vec()
+                    .into_iter()
+                    .map(|x| x.0)
+                    .collect::<Vec<i64>>()
+                    .into()
+            })
         }
     }
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn max_n_int_trans(
     state: Internal,
     value: i64,
@@ -63,10 +59,10 @@ pub fn max_n_int_trans(
     .internal()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn max_n_int_rollup_trans(
     state: Internal,
-    value: toolkit_experimental::MaxInts<'static>,
+    value: MaxInts<'static>,
     fcinfo: pg_sys::FunctionCallInfo,
 ) -> Option<Internal> {
     let values: Vec<Reverse<i64>> = value.values.clone().into_iter().map(Reverse).collect();
@@ -79,7 +75,7 @@ pub fn max_n_int_rollup_trans(
     .internal()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn max_n_int_combine(state1: Internal, state2: Internal) -> Option<Internal> {
     nmost_trans_combine(unsafe { state1.to_inner::<MaxIntTransType>() }, unsafe {
         state2.to_inner::<MaxIntTransType>()
@@ -87,55 +83,62 @@ pub fn max_n_int_combine(state1: Internal, state2: Internal) -> Option<Internal>
     .internal()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn max_n_int_serialize(state: Internal) -> bytea {
     let state: Inner<MaxIntTransType> = unsafe { state.to_inner().unwrap() };
     crate::do_serialize!(state)
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn max_n_int_deserialize(bytes: bytea, _internal: Internal) -> Option<Internal> {
     let i: MaxIntTransType = crate::do_deserialize!(bytes, MaxIntTransType);
     Internal::new(i).into()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
-pub fn max_n_int_final(state: Internal) -> toolkit_experimental::MaxInts<'static> {
+#[pg_extern(immutable, parallel_safe)]
+pub fn max_n_int_final(state: Internal) -> MaxInts<'static> {
     unsafe { &mut *state.to_inner::<MaxIntTransType>().unwrap() }.into()
 }
 
-#[pg_extern(
-    schema = "toolkit_experimental",
-    name = "into_array",
-    immutable,
-    parallel_safe
-)]
-pub fn max_n_int_to_array(agg: toolkit_experimental::MaxInts<'static>) -> Vec<i64> {
+#[pg_extern(name = "into_array", immutable, parallel_safe)]
+pub fn max_n_int_to_array(agg: MaxInts<'static>) -> Vec<i64> {
     agg.values.clone().into_vec()
 }
 
-#[pg_extern(
-    schema = "toolkit_experimental",
-    name = "into_values",
-    immutable,
-    parallel_safe
-)]
-pub fn max_n_int_to_values(agg: toolkit_experimental::MaxInts<'static>) -> SetOfIterator<i64> {
+#[pg_extern(name = "into_values", immutable, parallel_safe)]
+pub fn max_n_int_to_values(agg: MaxInts<'static>) -> SetOfIterator<i64> {
     SetOfIterator::new(agg.values.clone().into_iter())
+}
+
+#[pg_operator(immutable, parallel_safe)]
+#[opname(->)]
+pub fn arrow_max_int_into_values<'a>(
+    agg: MaxInts<'static>,
+    _accessor: AccessorIntoValues<'a>,
+) -> SetOfIterator<'a, i64> {
+    max_n_int_to_values(agg)
+}
+#[pg_operator(immutable, parallel_safe)]
+#[opname(->)]
+pub fn arrow_max_int_into_array<'a>(
+    agg: MaxInts<'static>,
+    _accessor: AccessorIntoArray<'a>,
+) -> Vec<i64> {
+    max_n_int_to_array(agg)
 }
 
 extension_sql!(
     "\n\
-    CREATE AGGREGATE toolkit_experimental.max_n(\n\
+    CREATE AGGREGATE max_n(\n\
         value bigint, capacity bigint\n\
     ) (\n\
-        sfunc = toolkit_experimental.max_n_int_trans,\n\
+        sfunc = max_n_int_trans,\n\
         stype = internal,\n\
-        combinefunc = toolkit_experimental.max_n_int_combine,\n\
+        combinefunc = max_n_int_combine,\n\
         parallel = safe,\n\
-        serialfunc = toolkit_experimental.max_n_int_serialize,\n\
-        deserialfunc = toolkit_experimental.max_n_int_deserialize,\n\
-        finalfunc = toolkit_experimental.max_n_int_final\n\
+        serialfunc = max_n_int_serialize,\n\
+        deserialfunc = max_n_int_deserialize,\n\
+        finalfunc = max_n_int_final\n\
     );\n\
 ",
     name = "max_n_int",
@@ -150,16 +153,16 @@ extension_sql!(
 
 extension_sql!(
     "\n\
-    CREATE AGGREGATE toolkit_experimental.rollup(\n\
-        value toolkit_experimental.MaxInts\n\
+    CREATE AGGREGATE rollup(\n\
+        value MaxInts\n\
     ) (\n\
-        sfunc = toolkit_experimental.max_n_int_rollup_trans,\n\
+        sfunc = max_n_int_rollup_trans,\n\
         stype = internal,\n\
-        combinefunc = toolkit_experimental.max_n_int_combine,\n\
+        combinefunc = max_n_int_combine,\n\
         parallel = safe,\n\
-        serialfunc = toolkit_experimental.max_n_int_serialize,\n\
-        deserialfunc = toolkit_experimental.max_n_int_deserialize,\n\
-        finalfunc = toolkit_experimental.max_n_int_final\n\
+        serialfunc = max_n_int_serialize,\n\
+        deserialfunc = max_n_int_deserialize,\n\
+        finalfunc = max_n_int_final\n\
     );\n\
 ",
     name = "max_n_int_rollup",
@@ -199,17 +202,40 @@ mod tests {
             }
 
             // Test into_array
-            let result =
-                client.update("SELECT toolkit_experimental.into_array(toolkit_experimental.max_n(val, 5)) from data",
-                    None, None,
-                ).unwrap().first().get_one::<Vec<i64>>().unwrap();
+            let result = client
+                .update("SELECT into_array(max_n(val, 5)) from data", None, None)
+                .unwrap()
+                .first()
+                .get_one::<Vec<i64>>()
+                .unwrap();
+            assert_eq!(result.unwrap(), vec![99, 98, 97, 96, 95]);
+            let result = client
+                .update("SELECT max_n(val, 5)->into_array() from data", None, None)
+                .unwrap()
+                .first()
+                .get_one::<Vec<i64>>()
+                .unwrap();
             assert_eq!(result.unwrap(), vec![99, 98, 97, 96, 95]);
 
             // Test into_values
-            let mut result =
-                client.update("SELECT toolkit_experimental.into_values(toolkit_experimental.max_n(val, 3))::TEXT from data",
-                    None, None,
-                ).unwrap();
+            let mut result = client
+                .update(
+                    "SELECT into_values(max_n(val, 3))::TEXT from data",
+                    None,
+                    None,
+                )
+                .unwrap();
+            assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("99"));
+            assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("98"));
+            assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("97"));
+            assert!(result.next().is_none());
+            let mut result = client
+                .update(
+                    "SELECT (max_n(val, 3)->into_values())::TEXT from data",
+                    None,
+                    None,
+                )
+                .unwrap();
             assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("99"));
             assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("98"));
             assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("97"));
@@ -218,8 +244,8 @@ mod tests {
             // Test rollup
             let result =
                 client.update(
-                    "WITH aggs as (SELECT category, toolkit_experimental.max_n(val, 5) as agg from data GROUP BY category)
-                        SELECT toolkit_experimental.into_array(toolkit_experimental.rollup(agg)) FROM aggs",
+                    "WITH aggs as (SELECT category, max_n(val, 5) as agg from data GROUP BY category)
+                        SELECT into_array(rollup(agg)) FROM aggs",
                         None, None,
                     ).unwrap().first().get_one::<Vec<i64>>().unwrap();
             assert_eq!(result.unwrap(), vec![99, 98, 97, 96, 95]);