+++ extension/src/nmost/min_int.rs
@@ -3,6 +3,7 @@ use pgx::{iter::SetOfIterator, *};
 use crate::nmost::*;
 
 use crate::{
+    accessors::{AccessorIntoArray, AccessorIntoValues},
     flatten,
     palloc::{Inner, Internal, InternalAsValue, ToInternal},
     pg_type,
@@ -12,35 +13,30 @@ use crate::{
 
 type MinIntTransType = NMostTransState<i64>;
 
-#[pg_schema]
-pub mod toolkit_experimental {
-    use super::*;
-
-    pg_type! {
-        #[derive(Debug)]
-        struct MinInts <'input> {
-            capacity : u32,
-            elements : u32,
-            values : [i64; self.elements],
-        }
+pg_type! {
+    #[derive(Debug)]
+    struct MinInts <'input> {
+        capacity : u32,
+        elements : u32,
+        values : [i64; self.elements],
     }
-    ron_inout_funcs!(MinInts);
+}
+ron_inout_funcs!(MinInts);
 
-    impl<'input> From<&mut MinIntTransType> for MinInts<'input> {
-        fn from(item: &mut MinIntTransType) -> Self {
-            let heap = std::mem::take(&mut item.heap);
-            unsafe {
-                flatten!(MinInts {
-                    capacity: item.capacity as u32,
-                    elements: heap.len() as u32,
-                    values: heap.into_sorted_vec().into()
-                })
-            }
+impl<'input> From<&mut MinIntTransType> for MinInts<'input> {
+    fn from(item: &mut MinIntTransType) -> Self {
+        let heap = std::mem::take(&mut item.heap);
+        unsafe {
+            flatten!(MinInts {
+                capacity: item.capacity as u32,
+                elements: heap.len() as u32,
+                values: heap.into_sorted_vec().into()
+            })
         }
     }
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn min_n_int_trans(
     state: Internal,
     value: i64,
@@ -56,10 +52,10 @@ pub fn min_n_int_trans(
     .internal()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn min_n_int_rollup_trans(
     state: Internal,
-    value: toolkit_experimental::MinInts<'static>,
+    value: MinInts<'static>,
     fcinfo: pg_sys::FunctionCallInfo,
 ) -> Option<Internal> {
     nmost_rollup_trans_function(
@@ -71,7 +67,7 @@ pub fn min_n_int_rollup_trans(
     .internal()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn min_n_int_combine(state1: Internal, state2: Internal) -> Option<Internal> {
     nmost_trans_combine(unsafe { state1.to_inner::<MinIntTransType>() }, unsafe {
         state2.to_inner::<MinIntTransType>()
@@ -79,55 +75,62 @@ pub fn min_n_int_combine(state1: Internal, state2: Internal) -> Option<Internal>
     .internal()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn min_n_int_serialize(state: Internal) -> bytea {
     let state: Inner<MinIntTransType> = unsafe { state.to_inner().unwrap() };
     crate::do_serialize!(state)
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
+#[pg_extern(immutable, parallel_safe)]
 pub fn min_n_int_deserialize(bytes: bytea, _internal: Internal) -> Option<Internal> {
     let i: MinIntTransType = crate::do_deserialize!(bytes, MinIntTransType);
     Internal::new(i).into()
 }
 
-#[pg_extern(schema = "toolkit_experimental", immutable, parallel_safe)]
-pub fn min_n_int_final(state: Internal) -> toolkit_experimental::MinInts<'static> {
+#[pg_extern(immutable, parallel_safe)]
+pub fn min_n_int_final(state: Internal) -> MinInts<'static> {
     unsafe { &mut *state.to_inner::<MinIntTransType>().unwrap() }.into()
 }
 
-#[pg_extern(
-    schema = "toolkit_experimental",
-    name = "into_array",
-    immutable,
-    parallel_safe
-)]
-pub fn min_n_int_to_array(agg: toolkit_experimental::MinInts<'static>) -> Vec<i64> {
+#[pg_extern(name = "into_array", immutable, parallel_safe)]
+pub fn min_n_int_to_array(agg: MinInts<'static>) -> Vec<i64> {
     agg.values.clone().into_vec()
 }
 
-#[pg_extern(
-    schema = "toolkit_experimental",
-    name = "into_values",
-    immutable,
-    parallel_safe
-)]
-pub fn min_n_int_to_values(agg: toolkit_experimental::MinInts<'static>) -> SetOfIterator<i64> {
+#[pg_extern(name = "into_values", immutable, parallel_safe)]
+pub fn min_n_int_to_values(agg: MinInts<'static>) -> SetOfIterator<i64> {
     SetOfIterator::new(agg.values.clone().into_iter())
+}
+
+#[pg_operator(immutable, parallel_safe)]
+#[opname(->)]
+pub fn arrow_min_int_into_values<'a>(
+    agg: MinInts<'static>,
+    _accessor: AccessorIntoValues<'a>,
+) -> SetOfIterator<'a, i64> {
+    min_n_int_to_values(agg)
+}
+#[pg_operator(immutable, parallel_safe)]
+#[opname(->)]
+pub fn arrow_min_int_into_array<'a>(
+    agg: MinInts<'static>,
+    _accessor: AccessorIntoArray<'a>,
+) -> Vec<i64> {
+    min_n_int_to_array(agg)
 }
 
 extension_sql!(
     "\n\
-    CREATE AGGREGATE toolkit_experimental.min_n(\n\
+    CREATE AGGREGATE min_n(\n\
         value bigint, capacity bigint\n\
     ) (\n\
-        sfunc = toolkit_experimental.min_n_int_trans,\n\
+        sfunc = min_n_int_trans,\n\
         stype = internal,\n\
-        combinefunc = toolkit_experimental.min_n_int_combine,\n\
+        combinefunc = min_n_int_combine,\n\
         parallel = safe,\n\
-        serialfunc = toolkit_experimental.min_n_int_serialize,\n\
-        deserialfunc = toolkit_experimental.min_n_int_deserialize,\n\
-        finalfunc = toolkit_experimental.min_n_int_final\n\
+        serialfunc = min_n_int_serialize,\n\
+        deserialfunc = min_n_int_deserialize,\n\
+        finalfunc = min_n_int_final\n\
     );\n\
 ",
     name = "min_n_int",
@@ -142,16 +145,16 @@ extension_sql!(
 
 extension_sql!(
     "\n\
-    CREATE AGGREGATE toolkit_experimental.rollup(\n\
-        value toolkit_experimental.MinInts\n\
+    CREATE AGGREGATE rollup(\n\
+        value MinInts\n\
     ) (\n\
-        sfunc = toolkit_experimental.min_n_int_rollup_trans,\n\
+        sfunc = min_n_int_rollup_trans,\n\
         stype = internal,\n\
-        combinefunc = toolkit_experimental.min_n_int_combine,\n\
+        combinefunc = min_n_int_combine,\n\
         parallel = safe,\n\
-        serialfunc = toolkit_experimental.min_n_int_serialize,\n\
-        deserialfunc = toolkit_experimental.min_n_int_deserialize,\n\
-        finalfunc = toolkit_experimental.min_n_int_final\n\
+        serialfunc = min_n_int_serialize,\n\
+        deserialfunc = min_n_int_deserialize,\n\
+        finalfunc = min_n_int_final\n\
     );\n\
 ",
     name = "min_n_int_rollup",
@@ -191,17 +194,40 @@ mod tests {
             }
 
             // Test into_array
-            let result =
-                client.update("SELECT toolkit_experimental.into_array(toolkit_experimental.min_n(val, 5)) from data",
-                    None, None,
-                ).unwrap().first().get_one::<Vec<i64>>().unwrap();
+            let result = client
+                .update("SELECT into_array(min_n(val, 5)) from data", None, None)
+                .unwrap()
+                .first()
+                .get_one::<Vec<i64>>()
+                .unwrap();
+            assert_eq!(result.unwrap(), vec![0, 1, 2, 3, 4]);
+            let result = client
+                .update("SELECT min_n(val, 5)->into_array() from data", None, None)
+                .unwrap()
+                .first()
+                .get_one::<Vec<i64>>()
+                .unwrap();
             assert_eq!(result.unwrap(), vec![0, 1, 2, 3, 4]);
 
             // Test into_values
-            let mut result =
-                client.update("SELECT toolkit_experimental.into_values(toolkit_experimental.min_n(val, 3))::TEXT from data",
-                    None, None,
-                ).unwrap();
+            let mut result = client
+                .update(
+                    "SELECT into_values(min_n(val, 3))::TEXT from data",
+                    None,
+                    None,
+                )
+                .unwrap();
+            assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("0"));
+            assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("1"));
+            assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("2"));
+            assert!(result.next().is_none());
+            let mut result = client
+                .update(
+                    "SELECT (min_n(val, 3)->into_values())::TEXT from data",
+                    None,
+                    None,
+                )
+                .unwrap();
             assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("0"));
             assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("1"));
             assert_eq!(result.next().unwrap()[1].value().unwrap(), Some("2"));
@@ -210,8 +236,8 @@ mod tests {
             // Test rollup
             let result =
                 client.update(
-                    "WITH aggs as (SELECT category, toolkit_experimental.min_n(val, 5) as agg from data GROUP BY category)
-                        SELECT toolkit_experimental.into_array(toolkit_experimental.rollup(agg)) FROM aggs",
+                    "WITH aggs as (SELECT category, min_n(val, 5) as agg from data GROUP BY category)
+                        SELECT into_array(rollup(agg)) FROM aggs",
                         None, None,
                     ).unwrap().first().get_one::<Vec<i64>>().unwrap();
             assert_eq!(result.unwrap(), vec![0, 1, 2, 3, 4]);